refactor(phase-3B-2): split api_state.py into 7 sibling viz modules [no behavior change]
Phase 3B PR-3B-2 / viz/REFACTORING.md §2 Step 2.

``ari/viz/api_state.py`` shrinks from 1404 → 76 lines (re-export
facade only).  Per-cluster bodies now live in:

- ``checkpoint_finder.py``     — discovery + PID liveness probe (57 L).
- ``state_sync.py``            — node-tree loading + WebSocket
  broadcast + watcher thread (118 L).
- ``checkpoint_api.py``        — model list / checkpoint list /
  summary / lineage decisions (272 L).
- ``ear.py``                   — EAR curate/publish/clone helpers,
  per-node report endpoint (430 L).
- ``file_api.py``              — paper dir + file read/save/upload/
  delete + LaTeX compile (296 L).
- ``checkpoint_lifecycle.py``  — checkpoint delete + switch (193 L).
- ``node_work_api.py``         — per-node filetree / filecontent /
  memory listing (210 L).

api_state.py becomes a thin facade re-exporting every public symbol.
Downstream callers (``server.py`` route table, GUI tests, etc.) keep
their ``from .api_state import _api_X`` imports unchanged.

Test-monkeypatch compatibility (the heart of the PR):
- Cross-module helpers used inside split modules look up the helper
  through ``api_state`` at call time:
  - ``checkpoint_finder._resolve_checkpoint_dir`` reads
    ``api_state._checkpoint_search_bases`` so tests that swap it
    flow through.
  - ``state_sync._broadcast`` / ``_watcher_thread`` defer
    ``_do_broadcast`` / ``_load_nodes_tree`` lookups to ``api_state``.
  - Each split module exposes ``_resolve_checkpoint_dir`` (and the
    other shared helpers) as a thin wrapper that fetches the current
    binding from ``api_state``, so tests using
    ``patch.object(api_state, "_resolve_checkpoint_dir", ...)`` are
    intercepted here.

Module-level constants restored to their owning files:
- ``file_api.py``: ``PAPER_DIR_NAME`` / ``_TEXT_EXTENSIONS`` /
  ``_PAPER_ROOT_ARTEFACTS`` / ``_FIGURE_GLOBS``.
- ``ear.py``:      ``_PUBLISH_YAML_DEFAULT``.
- ``node_work_api.py``: ``_BINARY_EXTENSIONS`` / ``_SKIP_DIRS``.

Test fixture update:
- ``test_workflow_contract.py:test_api_state_still_reads_full_paper_tex``
  concatenates every ``viz/*.py`` so the ``"full_paper.tex"`` literal
  search succeeds regardless of which file owns it.

pytest ari-core/tests/ -q (excluding live integrations):
1971 passed, 15 skipped (unchanged).
